ExitTiles, Early Win System, and Reworking Placing Walls Part 1 - Added exits, when the player enters the exit tile, they win the level. Multiple types, doors being able to be locked, not yet fully implemented/tested. - Reworked several parts of the models/views/controllers accordingly to allow for winning, saving/loading/placing/storing reference to exit tiles, creating textures from text, etc. - Began reworking placing walls in the level editor to a much more simplistic and automatic system.

+++ Game Engine/Enemy.cpp
@@ -70,7 +70,7 @@ void Enemy::determineMovement(double playerPosX, double playerPosY, std::vector<
 		{
 			float vectorX = playerPosX - getCenterPosX();
 			float vectorY = playerPosY - getCenterPosY();
-			double magnitude = sqrt(pow(vectorX, 2) + pow(vectorY, 2));
+			float magnitude = sqrt(pow(vectorX, 2) + pow(vectorY, 2));
 			//get unit vector:
 			vectorX /= magnitude;
 			vectorY /= magnitude;
